+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.10.1
+
+- add `listChunk` list utility
+
 # 0.10.0
 
 - export `pedantic` and `meta` in `common_utils_import.dart`